# docs/api/PRD-003-openapi.yaml
...
openapi: 3.1.0
info:
  title: Clean Connect API
  version: '1.0.5-prd003-build-ticket-3.3-photos-presign'
  description: |
    API publique Clean Connect — bloc **Photos & Paiements** (PRD-003).

    ## Conventions globales

    - **Versionnement URL** : `/v1/...` (NestJS `app.enableVersioning({ type: URI })`).
    - **Authentification** : JWT Bearer (`Authorization: Bearer <accessToken>`) sauf endpoints
      explicitement marqués `security: []` (webhook Stripe).
    - **Idempotency-Key** : header obligatoire sur toute opération financière ou mutation
      Stripe-bound. Format : `[A-Za-z0-9_-]{8,255}`. **Aucun trim** côté serveur — la clé
      est consommée *telle quelle*. Voir `IdempotencyKeyHeader`.
    - **Rétention idempotence (CTO)** : réponse identique garantie **au minimum 24 h** pour
      rejouer la même clé avec le même corps (remboursements, captures, presign, etc.).
      Même clé + corps différent avant expiration de la fenêtre → `409 PAYMENT_IDEMPOTENCY_CONFLICT`
      (paiements) ou équivalent domaine.
    - **Rate limiting** : annoté par endpoint via `x-rate-limit` (NestJS `@nestjs/throttler`).
      Réponse `429 Too Many Requests` standardisée.
    - **Erreurs métier stables** : chaque domaine expose un enum `*ErrorCode` versionné
      (cf. `PaymentErrorCode`, `PhotoErrorCode`, `WebhookErrorCode`, `AutoReleaseErrorCode`).
      Format de réponse erreur : `{ error: <code>, reason?: <string> }`.
...

    ## Sécurité — règles dures

...

...

...
    |---|---|---|---|
...

...

...

...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT d'accès (15 min). Issu de `POST /v1/auth/login` ou `POST /v1/auth/refresh`.
        Header : `Authorization: Bearer <accessToken>`.
    stripeSignature:
      type: apiKey
      in: header
      name: stripe-signature
      description: |
        Signature HMAC Stripe vérifiée via `stripe.webhooks.constructEvent()`.
        **Obligatoire** sur `POST /v1/webhooks/stripe`. Le webhook est rejeté en `400`
        si signature absente ou invalide (sans pollution de log). Réponse succès : `202`.

  # ---------------------------------------------------------------------------
  # Parameters réutilisables
  # ---------------------------------------------------------------------------
  parameters:
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 255
        pattern: '^[A-Za-z0-9_-]+$'
      description: |
        Clé d'idempotence client. Format : `[A-Za-z0-9_-]{8,255}`. **Aucun trim**.
        **Rétention minimale côté serveur : 24 h** (décision CTO) : rejouer la même clé
        avec le **même** corps dans cette fenêtre → **réponse identique** (cache idempotent,
        codes 200/201/202 selon l'endpoint). Même clé + corps **différent** dans la fenêtre
        → `409 PAYMENT_IDEMPOTENCY_CONFLICT` (paiements) ou conflit équivalent domaine.
      example: cc-mobile-7f8e9d1a-mission42
...
    UuidPathId:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
      description: UUID v4 de la ressource ciblée.
...
  responses:
    ValidationError:
      description: '400 — Erreur de validation (Zod / DTO).'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationErrorBody' }
    Unauthenticated:
      description: '401 — JWT absent / expiré / invalide.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/GenericErrorBody' }
    Forbidden:
      description: '403 — JWT valide mais rôle insuffisant ou ownership refusé.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/GenericErrorBody' }
    NotFound:
      description: '404 — Ressource introuvable.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/GenericErrorBody' }
...
    TooManyRequests:
      description: '429 — Rate limit dépassé (cf. `x-rate-limit` par endpoint).'
      headers:
        Retry-After:
          schema: { type: integer, description: 'Secondes avant nouvelle tentative.' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/GenericErrorBody' }
    UploadSessionExpired:
      description: '410 — Session d''upload expirée ou invalide (`UPLOAD_SESSION_EXPIRED`).'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/PhotoErrorBody' }

  # ---------------------------------------------------------------------------
  # Schemas — Public uniquement (RBAC-aware)
  # AUCUN schema *Internal ni *Raw n'est référencé ici (refus CTO).
  # Source : `@cc/shared-types/zod` (payment.ts, photo.ts, webhook.ts, etc.)
  # ---------------------------------------------------------------------------
  schemas:
...
    Uuid:
      type: string
      format: uuid
      example: 2a1f3b41-7e8a-4c5d-9b6f-1234567890ab
    IsoDateTime:
      type: string
      format: date-time
      example: '2026-05-12T09:42:18.123Z'
...
    PhotoType:
      type: string
      enum: [BEFORE, AFTER]
    PhotoVariant:
      type: string
      enum: [ORIGINAL, DISPLAY]
      description: |
        `DISPLAY` : variante optimisée, EXIF strip (côté client/prestataire).
        `ORIGINAL` : variante intègre sécurisée (admin only).
    PhotoMimeType:
      type: string
      enum: [image/jpeg, image/png, image/heic, image/webp]
...
    PhotoErrorCode:
      type: string
      enum:
        - PHOTO_NOT_FOUND
        - PHOTO_FORBIDDEN
        - PHOTO_INVALID_STATE
        - PHOTO_UPLOAD_SESSION_EXPIRED
        - PHOTO_UPLOAD_SESSION_ALREADY_CONSUMED
        - PHOTO_UPLOAD_SESSION_MISSION_MISMATCH
        - PHOTO_CAPTURE_CLIENT_UUID_SESSION_MISMATCH
        - PHOTO_CHECKSUM_MISMATCH
        - PHOTO_MIME_NOT_ALLOWED
        - PHOTO_MAX_BYTES_EXCEEDED
        - PHOTO_MAX_COUNT_EXCEEDED
        - PHOTO_INSUFFICIENT_BEFORE
        - PHOTO_INSUFFICIENT_AFTER
        - PHOTO_DELETION_FORBIDDEN
        - PHOTO_GPS_INCONSISTENT
        - UPLOAD_SESSION_EXPIRED
        - PHOTOS_DISABLED
...
    GenericErrorBody:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          # Référence informative — la valeur réelle dépend du domaine :
          # `PaymentErrorCode`, `PhotoErrorCode`, `WebhookErrorCode`, `AutoReleaseErrorCode`.
          type: string
          description: |
            Code erreur stable. Énumérés par domaine :
            - Paiements : voir `PaymentErrorCode`
            - Photos : voir `PhotoErrorCode`
            - Webhooks : voir `WebhookErrorCode`
            - Auto-release : voir `AutoReleaseErrorCode`
            - Workflow mission (start/complete) : voir `MissionWorkflowErrorCode`
            - Session upload expirée : `UPLOAD_SESSION_EXPIRED` (**HTTP 410**)
          examples:
            - PAYMENT_INVALID_STATE
            - PHOTO_UPLOAD_SESSION_EXPIRED
            - WEBHOOK_INVALID_SIGNATURE
            - AUTO_RELEASE_JOB_INVALID_STATE
        reason:
          type: string
          maxLength: 500
          description: Détail libre (jamais de PII).

    # Discriminated error bodies par domaine — référencés au moins une fois
    # pour exposer les enums `*ErrorCode` via les responses.
...
    PhotoErrorBody:
      allOf:
        - $ref: '#/components/schemas/GenericErrorBody'
        - type: object
          properties:
            error: { $ref: '#/components/schemas/PhotoErrorCode' }
...
    ValidationErrorBody:
      type: object
      required: [error, issues]
      additionalProperties: false
      properties:
        error:
          type: string
          enum: [VALIDATION_FAILED]
        issues:
          type: array
          items:
            type: object
            required: [path, code, message]
            additionalProperties: false
            properties:
              path: { type: array, items: { type: string } }
              code: { type: string }
              message: { type: string }

    # ----- Inputs : Payments ----------------------------------------------
...
    PhotoGpsInput:
      # Zod : `photoGpsInputSchema`
      type: object
      required: [gpsLatitude, gpsLongitude]
      additionalProperties: false
      properties:
        gpsLatitude:
          oneOf:
            - { type: number, minimum: -90, maximum: 90 }
            - { type: 'null' }
        gpsLongitude:
          oneOf:
            - { type: number, minimum: -180, maximum: 180 }
            - { type: 'null' }
        gpsAccuracyMeters:
          oneOf:
            - { type: integer, minimum: 0, maximum: 10000 }
            - { type: 'null' }
      description: |
        `gpsLatitude` et `gpsLongitude` doivent être **tous deux** renseignés ou
        **tous deux** NULL (cohérence vérifiée serveur — code `PHOTO_GPS_INCONSISTENT`).
        `gpsAccuracyMeters` interdit si lat/lng sont NULL.
    PresignPhotoUploadInput:
      # Zod : `presignPhotoUploadInputSchema`
      type: object
      required: [missionId, phase, variant, captureClientUuid, bytes, mimeType, gps]
      additionalProperties: false
      properties:
        missionId:
          allOf: [{ $ref: '#/components/schemas/Uuid' }]
          description: |
            **Doit strictement égaler** l'`{id}` du path (revue CTO Ticket 3.3 — anti
            redirection cross-mission depuis le body). Sinon → `409 PHOTO_UPLOAD_SESSION_MISSION_MISMATCH`.
        phase: { $ref: '#/components/schemas/PhotoType' }
        variant: { $ref: '#/components/schemas/PhotoVariant' }
        captureClientUuid:
          allOf: [{ $ref: '#/components/schemas/Uuid' }]
          description: |
            UUID v4 généré côté mobile. **Obligatoire au presign** (revue CTO).
            Stocké côté serveur sur la `PhotoUploadSession` (lien logique capture ↔ session).
            Sert à apparier ORIGINAL/DISPLAY et à garantir l'idempotence d'upload.
        bytes:
          type: integer
          minimum: 1
          maximum: 10485760
          description: Taille en bytes (max 10 Mo, ADR-009).
        mimeType: { $ref: '#/components/schemas/PhotoMimeType' }
        gps: { $ref: '#/components/schemas/PhotoGpsInput' }
    ConfirmPhotoUploadInput:
      # Zod : `confirmPhotoUploadInputSchema`
      type: object
      required:
        - photoUploadSessionId
        - sessionToken
        - photoId
        - captureClientUuid
        - cloudinaryPublicId
        - checksumSha256
        - imageWidth
        - imageHeight
        - bytes
        - mimeType
        - gps
      additionalProperties: false
      properties:
        photoUploadSessionId: { $ref: '#/components/schemas/Uuid' }
        sessionToken:
          type: string
          minLength: 32
          maxLength: 128
          description: |
            **Bearer-équivalent** de la session d'upload (PRD-003 Ticket 3.3).
            Le mobile renvoie le **token clair** reçu une seule fois au presign.
            Le serveur recalcule `sha256(sessionToken)` et compare au `tokenDigest`
            stocké en DB pour authentifier la session — protège contre les attaques
            par énumération d'`id` de session (anti session-id leak).

            **Sécurité** :
            - Jamais loggé (Pino redactor `*.sessionToken`).
            - Pas de stockage côté serveur (digest only).
            - Mismatch → `409 PHOTO_UPLOAD_SESSION_ALREADY_CONSUMED` (sémantique
              « session non trouvée », sans révéler si l'`id` existait ou non).
        photoId:
          allOf: [{ $ref: '#/components/schemas/Uuid' }]
          description: |
            **UUID v4 généré côté mobile** (offline-first PRD-003 Ticket 3.3).
            Le mobile DOIT réutiliser le **même** `photoId` à chaque retry après
            un échec réseau : l'idempotence forte est garantie côté DB via
            l'UNIQUE `(missionId, captureClientUuid, variant)` — pas via `photoId`.
            Si la même session a déjà produit une `Photo` pour ce `captureClientUuid`,
            la réponse renvoie `idempotent: true` avec le `photoId` initial.
        captureClientUuid:
          allOf: [{ $ref: '#/components/schemas/Uuid' }]
          description: |
            Doit **strictement** égaler le `captureClientUuid` stocké sur la
            `PhotoUploadSession` identifiée par `photoUploadSessionId` (revue CTO).
            Croisé avec `{id}` mission dans l'URL : empêche upload **cross-session** et
            **cross-mission**. Sinon → `409 PHOTO_CAPTURE_CLIENT_UUID_SESSION_MISMATCH`.
        cloudinaryPublicId:
          type: string
          minLength: 1
          maxLength: 1024
          description: |
            Doit **strictement** égaler le `cloudinaryPublicId` figé serveur au
            presign (la session scelle le path Cloudinary). Mismatch → `422`.
        checksumSha256:
          type: string
          minLength: 64
          maxLength: 64
          pattern: '^[0-9a-f]{64}$'
        imageWidth: { type: integer, minimum: 1, maximum: 20000 }
        imageHeight: { type: integer, minimum: 1, maximum: 20000 }
        bytes:
          type: integer
          minimum: 1
          maximum: 10485760
        mimeType: { $ref: '#/components/schemas/PhotoMimeType' }
        gps: { $ref: '#/components/schemas/PhotoGpsInput' }
    ConfirmPhotoUploadResponse:
      # Zod : `confirmPhotoUploadResponseSchema`
      type: object
      required:
        - photoId
        - missionId
        - phase
        - variant
        - captureClientUuid
        - bytes
        - imageWidth
        - imageHeight
        - gpsMissing
        - syncedAt
        - idempotent
      additionalProperties: false
      properties:
        photoId:
          allOf: [{ $ref: '#/components/schemas/Uuid' }]
          description: |
            ID de la `Photo` persistée. Si `idempotent: true`, c'est l'ID de la
            photo initiale (le `photoId` du body est ignoré au profit du premier).
        missionId: { $ref: '#/components/schemas/Uuid' }
        phase: { $ref: '#/components/schemas/PhotoType' }
        variant: { $ref: '#/components/schemas/PhotoVariant' }
        captureClientUuid: { $ref: '#/components/schemas/Uuid' }
        bytes:
          type: integer
          minimum: 1
          maximum: 10485760
          description: |
            Source de vérité **Cloudinary** (`getResource`), pas le report mobile.
        imageWidth: { type: integer, minimum: 1, maximum: 20000 }
        imageHeight: { type: integer, minimum: 1, maximum: 20000 }
        gpsMissing:
          type: boolean
          description: |
            `true` si GPS non fourni (lat+lng tous deux NULL au presign).
            Audit interne — mission **non bloquée** (décision CTO Q3).
        syncedAt: { $ref: '#/components/schemas/IsoDateTime' }
        idempotent:
          type: boolean
          description: |
            - `false` : première confirmation (création `Photo`).
            - `true` : replay confirm avec **mêmes** `(missionId, captureClientUuid, variant)` →
              aucun side-effect, on renvoie la `Photo` initiale.
            Sécurité offline-first : mobile peut rejouer `/confirm` après crash sans dupliquer.

    # ----- Public views : Photos ------------------------------------------
...
    PhotoUploadSignatureResponse:
      # Zod : `photoUploadSignatureResponseSchema`
      type: object
      required:
        - photoUploadSessionId
        - sessionToken
        - uploadUrl
        - cloudinaryParams
        - expiresAt
        - maxBytes
        - allowedMimeTypes
      additionalProperties: false
      properties:
        photoUploadSessionId: { $ref: '#/components/schemas/Uuid' }
        sessionToken:
          type: string
          minLength: 32
          maxLength: 128
          description: |
            **Token clair de la session — retourné UNE SEULE FOIS** (PRD-003 Ticket 3.3).

            - Généré par `crypto.randomBytes` côté serveur (base64-url, ≥ 256 bits d'entropie).
            - Le serveur stocke uniquement `sha256(sessionToken)` (`tokenDigest` UNIQUE en DB).
            - Le mobile DOIT le **conserver en mémoire** pour `POST /photos/confirm`.
            - **Non récupérable** : si le mobile le perd (kill app), la session est inutilisable
              jusqu'à expiration (`expiresAt`) — il faut rejouer le presign.
            - **Jamais loggé** (Pino redactor `*.sessionToken` + `*.tokenDigest`).

            Ce token authentifie la session au `/confirm` (mismatch → `409`).
        uploadUrl:
          type: string
          format: uri
          description: |
            URL Cloudinary direct (`https://api.cloudinary.com/v1_1/<cloud>/image/upload`).
            Le mobile upload directement en multipart, sans transiter par l'API.
        cloudinaryParams:
          type: object
          additionalProperties: { type: string }
          description: |
            Paramètres signés Cloudinary à intégrer au multipart (timestamp, signature,
            folder, public_id, type=private, etc.). **Confidentiel — jamais loggé**
            (filter Pino `*.cloudinaryParams.signature`, `*.api_key`).

            **Format livré (Ticket 3.3)** : `{ uploadUrl, cloudName, apiKey, publicId,
            folder, type, timestamp, signature, signatureAlgorithm, mimeType, maxBytes }`.
            `signatureAlgorithm` = `sha1` (Cloudinary SDK), pas `sha256`.
        expiresAt:
          allOf: [{ $ref: '#/components/schemas/IsoDateTime' }]
          description: |
            Expiration de la session (= `PHOTO_UPLOAD_SESSION_TTL_SECONDS`, défaut 300 s).
            Cloudinary applique sa propre validité côté signature multipart.
        maxBytes:
          type: integer
          minimum: 1
          maximum: 10485760
        allowedMimeTypes:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/PhotoMimeType' }
...
security:
  - bearerAuth: []

# =============================================================================
# Paths — 18 endpoints
# =============================================================================
paths:
...
  '/v1/missions/{id}/photos/presign':
    post:
      tags: [Photos]
      operationId: presignPhotoUpload
      summary: '[PRESTATAIRE] Demander une signed URL Cloudinary'
      description: |
        Émet une `PhotoUploadSession` à usage unique (expire en 5 min).
        Le mobile upload directement à Cloudinary (multipart) avec les
        `cloudinaryParams` retournés — **aucune photo ne transite par l'API**.

        **`captureClientUuid` (revue CTO)** :
        - **Obligatoire** dans le corps ; UUID v4 généré côté mobile.
        - Persisté sur la `PhotoUploadSession` : relie logiquement presign → confirm
          et apparie ORIGINAL/DISPLAY (même valeur pour les deux variants).
        - À `confirm`, le serveur vérifie que `captureClientUuid` du body **égale**
          celui de la session **et** que la session appartient à la mission `{id}` URL
          (anti cross-session / cross-mission).

        Sécurité :
        - `mimeType` validé contre whitelist (`image/jpeg|png|heic|webp`).
        - `bytes` ≤ 10 Mo. Cloudinary applique la même limite côté upload.
        - Ownership : prestataire **ACCEPTED** sur la mission `{id}`.

        Idempotence : session active existante pour `(missionId, captureClientUuid, variant)`
        → `409 PHOTO_UPLOAD_SESSION_ALREADY_CONSUMED`.
      x-rbac: [PRESTATAIRE]
      x-rate-limit: { strategy: 'token-bucket', max: 60, windowSeconds: 60, scope: 'user' }
      parameters:
        - $ref: '#/components/parameters/UuidPathId'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PresignPhotoUploadInput' }
      responses:
        '201':
          description: PhotoUploadSession créée.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PhotoUploadSignatureResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '409 — `PHOTO_UPLOAD_SESSION_ALREADY_CONSUMED` (idempotence).'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PhotoErrorBody' }
        '410': { $ref: '#/components/responses/UploadSessionExpired' }
        '422':
          description: '422 — règle métier (cf. `PhotoErrorCode`).'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PhotoErrorBody' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503':
          description: |
            `PHOTOS_DISABLED` — feature flag `FF_PHOTOS_ENABLED=false` côté serveur
            (PRD-003 Ticket 3.3). Module entièrement désactivé : aucune session créée.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PhotoErrorBody' }
  '/v1/missions/{id}/photos/confirm':
    post:
      tags: [Photos]
      operationId: confirmPhotoUpload
      summary: '[PRESTATAIRE] Confirmer un upload Cloudinary réussi'
      description: |
        Mobile signale que l'upload Cloudinary a réussi. Le serveur :
        1. Vérifie la `PhotoUploadSession` (non consommée ; si `expiresAt` dépassé → **`410 UPLOAD_SESSION_EXPIRED`**).
        2. Vérifie que le `{id}` mission URL = `missionId` de la session (**cross-mission** → `409`).
        3. Vérifie que `captureClientUuid` du body = celui stocké sur la session (**cross-session** / replay → `409 PHOTO_CAPTURE_CLIENT_UUID_SESSION_MISMATCH`).
        4. Vérifie via Cloudinary que l'asset existe + `checksumSha256` cohérent.
        5. Crée la ligne `Photo` ; marque la session `consumedAt = NOW()`.

        **Important (revue CTO)** : cet appel est **uniquement** « prestataire signale fin upload ».
        Il **ne** valide **pas** la mission côté client (cf. `POST /v1/missions/{id}/validate`).

        Sécurité :
        - `checksumSha256` recalculé / comparé serveur — mismatch → `409 PHOTO_CHECKSUM_MISMATCH`.
        - GPS : cohérence lat/lng (cf. `PhotoGpsInput`).
      x-rbac: [PRESTATAIRE]
      x-rate-limit: { strategy: 'token-bucket', max: 60, windowSeconds: 60, scope: 'user' }
      parameters:
        - $ref: '#/components/parameters/UuidPathId'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConfirmPhotoUploadInput' }
      responses:
        '201':
          description: |
            Photo enregistrée. `idempotent: false` = première confirmation ;
            `idempotent: true` = replay sans side-effect (mobile a rejoué après crash).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConfirmPhotoUploadResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: |
            `PHOTO_CHECKSUM_MISMATCH`, `PHOTO_UPLOAD_SESSION_ALREADY_CONSUMED`,
            `PHOTO_UPLOAD_SESSION_MISSION_MISMATCH`, `PHOTO_CAPTURE_CLIENT_UUID_SESSION_MISMATCH`.
            (Session expirée → **`410`** `UPLOAD_SESSION_EXPIRED`, pas 409.)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PhotoErrorBody' }
        '410': { $ref: '#/components/responses/UploadSessionExpired' }
        '422':
          description: |
            `PHOTO_INVALID_STATE` — règle métier non satisfaite (asset manquant côté
            Cloudinary, mismatch métadonnées width/height, etc.). Cf. `PhotoErrorCode`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PhotoErrorBody' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503':
          description: |
            `PHOTOS_DISABLED` — feature flag `FF_PHOTOS_ENABLED=false` côté serveur.
            Aucun upload accepté tant que le flag n'est pas activé (cf. PRD-003 Ticket 3.3).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PhotoErrorBody' }
...